requirements.txt
Flask==2.0.1
Pillow==10.0.0
requests==2.26.0
Werkzeug==2.0.3
flask-cors==3.0.10